Refactor function for get documents
