+++ app/src/main/java/com/jalasoft/routesapp/data/model/remote/City.kt
@@ -1,5 +1,7 @@
 package com.jalasoft.routesapp.data.model.remote
 
+import java.io.Serializable
+
 data class City(
     val name: String = "",
     val country: String = "",
@@ -7,4 +9,4 @@ data class City(
     val countryPhoneCode: String = "",
     val lat: String = "",
     val lng: String = ""
-)
+) : Serializable
