+++ app/src/main/java/com/jalasoft/routesapp/data/model/remote/Line.kt
@@ -1,9 +1,10 @@
 package com.jalasoft.routesapp.data.model.remote
 
 import com.google.firebase.firestore.GeoPoint
+import java.io.Serializable
 
 data class Line(
     val route: List<GeoPoint>,
     val start: List<String>,
     val stop: List<String>
-)
+) : Serializable
