+++ app/src/main/java/com/jalasoft/routesapp/data/model/remote/User.kt
@@ -1,5 +1,7 @@
 package com.jalasoft.routesapp.data.model.remote
 
+import java.io.Serializable
+
 data class User(
     val id: String? = null,
     val name: String? = null,
@@ -9,4 +11,4 @@ data class User(
     val typeLogin: Int? = null,
     val createdAt: Double? = null,
     val updatedAt: Double? = null
-)
+) : Serializable
