+++ app/src/main/java/com/jalasoft/routesapp/data/remote/managers/FirebaseManager.kt
@@ -7,7 +7,7 @@ import com.jalasoft.routesapp.util.Response
 import com.jalasoft.routesapp.util.helpers.FirebaseCollections
 import kotlinx.coroutines.tasks.await
 
-class FirebaseManager(private val db: FirebaseFirestore) : FirebaseDataSource {
+class FirebaseManager(val db: FirebaseFirestore) : FirebaseDataSource {
     fun getDocId(collection: FirebaseCollections): String {
         return db.collection(collection.toString()).document().id
     }
@@ -25,6 +25,15 @@ class FirebaseManager(private val db: FirebaseFirestore) : FirebaseDataSource {
         return try {
             val list = db.collection(collection.toString()).whereEqualTo(field, parameter).get().await()
             return Response.Success(list.toObjects(User::class.java))
+        } catch (e: Exception) {
+            Response.Error(e.message.toString(), null)
+        }
+    }
+
+    suspend inline fun <reified T : Any> getAllDocuments(collection: FirebaseCollections): Response<List<T>> {
+        return try {
+            val result = db.collection(collection.toString()).get().await()
+            Response.Success(result.toObjects(T::class.java))
         } catch (e: Exception) {
             Response.Error(e.message.toString(), null)
         }
