+++ app/src/main/java/com/jalasoft/routesapp/di/ManagersModule.kt
@@ -2,6 +2,7 @@ package com.jalasoft.routesapp.di
 
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
+import com.jalasoft.routesapp.data.remote.interfaces.CountryRepository
 import com.jalasoft.routesapp.data.remote.managers.*
 import dagger.Module
 import dagger.Provides
@@ -28,6 +29,6 @@ object ManagersModule {
     @Singleton
     @Provides
     fun provideCountryRepository(): CountryRepository {
-        return CountryManager()
+        return CountryManager(FirebaseManager(FirebaseFirestore.getInstance()))
     }
 }
